requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.0
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
execnet==2.1.1
fastapi==0.111.1
fastapi-cli==0.0.4
flake8==7.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
packaging==24.1
parse==1.20.2
parse-type==0.6.2
platformdirs==4.2.2
pluggy==1.5.0
psycopg2==2.9.9
pycodestyle==2.12.0
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.2.6
pytest==8.3.2
pytest-bdd==7.2.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
setuptools==72.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlmodel==0.0.21
starlette==0.37.2
structlog==24.4.0
testcontainers==4.7.2
tomlkit==0.13.0
typer==0.12.3
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
wrapt==1.16.0
